refactor: 승패 판정 클래스 분리
