+++ src/main/java/blackjack/Application.java
@@ -1,6 +1,7 @@
 package blackjack;
 
 import blackjack.controller.BlackjackController;
+import blackjack.domain.Referee;
 import blackjack.service.CardDistributor;
 import blackjack.service.RandomCardPicker;
 import blackjack.service.RandomNumberGenerator;
@@ -11,8 +12,9 @@ public class Application {
         RandomNumberGenerator randomNumberGenerator = new RandomNumberGenerator();
         RandomCardPicker cardPicker = new RandomCardPicker(randomNumberGenerator);
         CardDistributor cardDistributor = new CardDistributor(cardPicker);
+        Referee referee = new Referee();
 
-        BlackjackController blackjackController = new BlackjackController(cardDistributor);
+        BlackjackController blackjackController = new BlackjackController(cardDistributor, referee);
 
         blackjackController.startGame();
     }
